Inline post-message action runtime
