+++ README.md
@@ -25,4 +25,5 @@ The action always sends to `https://api.ro.am/v1/chat.post`.
 Notes:
 - `text` and `blocks` are mutually exclusive.
 - `blocks` is forwarded raw without action-side parsing or validation.
-- `chat-id` accepts legacy tagged IDs and maps them into the matching v1 destination field: `chatId`, `groupId`, or `userIds`.
+- `chat-id` accepts UUIDs and existing tagged IDs.
+- Tagged `chat-id` values are mapped to the matching v1 destination field: `chatId`, `groupId`, or `userIds`.
