+++ post-message/action.yml
@@ -6,13 +6,13 @@ inputs:
     description: Bearer token for the Roam API.
   chat-id:
     required: true
-    description: Existing destination identifier. Legacy tagged IDs are mapped automatically into the appropriate v1 destination field.
+    description: Destination identifier. Accepts UUIDs and existing tagged IDs.
   text:
     required: false
-    description: Text message content for v1 chat.post. Mutually exclusive with blocks.
+    description: Text message content. Mutually exclusive with blocks.
   blocks:
     required: false
-    description: Raw JSON for the blocks field. The action forwards this value to v1/chat.post without parsing.
+    description: Raw JSON for the blocks field. The action forwards this value without parsing.
   color:
     required: false
     description: Optional block message color. Only used when blocks is provided.
